Move getGappyRandomData to src/lib/metrics
This is a common function for returning gappy data for tests and is not
AWS specific

It should be in the common metrics code instead
