+++ src/lib/metrics/index.ts
@@ -1,4 +1,5 @@
 import moment from 'moment';
+import roundDown from '../moment/round';
 
 export type ElasticsearchMetricName = string;
 
@@ -48,3 +49,33 @@ export interface IMetricGraphDisplayable {
   readonly summaries: ReadonlyArray<IMetricSerieSummary>;
   readonly series: ReadonlyArray<IMetricSerie>;
 }
+
+/* tslint:disable:insecure-random */
+/* istanbul ignore next */
+export function getGappyRandomData(): {
+  readonly timestamps: ReadonlyArray<string>,
+  readonly values: ReadonlyArray<number>,
+} {
+  const minutesInADay = 24 * 60;
+  const timestamps: string[] = [];
+  const values: number[] = [];
+
+  const startTime = roundDown(moment().subtract(1, 'day'), moment.duration(5, 'minutes'));
+  for (let i = 0; i < minutesInADay; i += 5) {
+    if (i < minutesInADay * 0.2 || i > minutesInADay * 0.8 && i < minutesInADay * 0.9) {
+      // do nothing - empty piece of the graph
+    } else {
+      timestamps.push(startTime.clone().add(i, 'minutes').toISOString());
+      let value = 0;
+      if (values.length === 0) {
+        value = Math.random() * 105;
+      } else {
+        value = values[values.length - 1] + 10 * (Math.random() - 0.5);
+        value = value > 0 ? value : 0;
+      }
+      values.push(value);
+    }
+  }
+  return { timestamps, values };
+}
+/* tslint:enable:insecure-random */
